refactor: 갤러리 권한 로직 core:util로 이동
